fix(head): move fetchHeadContent to its own file
