+++ src/pwa/stores/connection/index.js
@@ -7,8 +7,6 @@ import {
   getEnv,
 } from 'mobx-state-tree';
 import { normalize } from 'normalizr';
-import { decode } from 'he';
-import getHeadContent from './head/getHeadContent';
 import { join } from './utils';
 import Entity from './entity';
 import entityShape from './entity-shape';
@@ -254,24 +252,5 @@ export const actions = self => {
         if (total.pages) list.total.pages = total.pages;
       }
     },
-    fetchHeadContent: flow(function* fetchHeadContent() {
-      try {
-        self.head.hasFailed = false;
-        const url = self.root.build.initialUrl;
-
-        if (!url) throw new Error('No initial url found.');
-
-        const { text: html } = yield getEnv(self).request(url);
-        const headHtml = html.match(
-          /<\s*?head[^>]*>([\w\W]+)<\s*?\/\s*?head\s*?>/,
-        )[1];
-        const { title, content } = getHeadContent(headHtml);
-
-        self.head.title = decode(title);
-        self.head.content = content;
-      } catch (error) {
-        self.head.hasFailed = true;
-      }
-    }),
   };
 };
